apps/flask-worker/app/requirements.txt
flask
gunicorn
python-dotenv
kafka-python
psycopg2-binary
pytz
schedule
requests